ocr-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
httpx==0.27.2
pymupdf==1.24.14
pillow==11.0.0
paddleocr==2.7.3
paddlepaddle==2.6.2
